% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Perform bubble sorting, counting the number of steps.}
\usage{
bubble_sort(data)
}
\arguments{
\item{data}{A list of vector and integer.
For example, in R code, data = list(vec = c(1, 0, 1, 1, 0, ...), cnt = 0).
The integer (cnt = 0) is the initial number of steps for bubble sorting.}
}
\description{
Perform bubble sorting, counting the number of steps.
}
