% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signs.R
\name{cluster_genesets}
\alias{cluster_genesets}
\title{Cluster each functional gene set into three groups.}
\usage{
cluster_genesets(sce, cormat, th_posi, th_nega)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{cormat}{A correlation matrix of gene expressions.}

\item{th_posi}{A threshold of positive correlation coefficient.}

\item{th_nega}{A threshold of negative correlation coefficient.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function clusters each functional gene set into strongly, variably, and
  weakly correlated gene sets.
}
