% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{collect_CO_asrt}
\alias{collect_CO_asrt}
\title{Collect a Cell Ontology database.}
\usage{
collect_CO_asrt(orgdb)
}
\arguments{
\item{orgdb}{A genome annotation package.}
}
\value{
Results from getCellOnto().
}
\description{
This function performs cellTypeToGenes() for collecting a Cell Ontology
  database.
}
