% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{collect_GO_asrt}
\alias{collect_GO_asrt}
\title{Collect Gene Ontology database.}
\usage{
collect_GO_asrt(orgdb)
}
\arguments{
\item{orgdb}{A genome annotation package.}
}
\value{
Results from groupGO().
}
\description{
This function collects Gene Ontology database.
}
