% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{collect_KEGG}
\alias{collect_KEGG}
\title{Collect KEGG database.}
\usage{
collect_KEGG(organism, categories)
}
\arguments{
\item{organism}{An identifier of organism.}

\item{categories}{Category name.}
}
\value{
Results from keggGet().
}
\description{
This function collects KEGG database.
}
