% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_signs.R
\name{compute_nswap_001}
\alias{compute_nswap_001}
\title{Compute number of swaps.}
\usage{
compute_nswap_001(enco)
}
\arguments{
\item{enco}{A vector.}
}
\description{
This function computes the number of swaps.
}
