% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_signs.R
\name{compute_nswaps}
\alias{compute_nswaps}
\title{Compute an edit distance between two vectors.}
\usage{
compute_nswaps(vec1, vec2)
}
\arguments{
\item{vec1}{A first vector.}

\item{vec2}{A second vector.}
}
\value{
An integer representing an edit distance.
}
\description{
This function computes the number of swaps of adjacent elements required for
  transforming vector 1 to vector 2, which have the same length and same
  amount of given numbers (e.g., vector1 = c(0,1,1), vector2 = c(1,1,0)).
  The following R code is inspired by Kendall tau distance algorithm.
}
