% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_signs.R
\name{compute_sepI_all}
\alias{compute_sepI_all}
\title{Compute separation indices for each cluster against the others.}
\usage{
compute_sepI_all(sce, labels)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{labels}{A vector of labels of all the samples (cells).}
}
\value{
A SingleCellExperiment object.
}
\description{
This function computes separation indices for each cluster versus the others.
}
