% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signs.R
\name{create_sce_signmatrix}
\alias{create_sce_signmatrix}
\title{Create a new SingleCellExperiment object for sign-by-sample matrices.}
\usage{
create_sce_signmatrix(sce, weight_strg = 0.5, weight_vari = 0.5)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{weight_strg}{A weight parameter for strongly correlated gene sets.}

\item{weight_vari}{A weight parameter for variably correlated gene sets.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function creates a new SingleCellExperiment object for sign-by-sample
  matrices (SSM) by concatenating SSMs for strongly and variably correlated
  gene sets.
}
