% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signs.R
\name{create_signs}
\alias{create_signs}
\title{Define signs for strongly and variably correlated gene sets.}
\usage{
create_signs(sce, min_cnt_strg = 2, min_cnt_vari = 2)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{min_cnt_strg}{An integer for the cutoff value for strongly correlated
gene sets.}

\item{min_cnt_vari}{An integer for the cutoff value for variably correlated
gene sets.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function define signs for strongly and variably correlated gene sets.
}
