% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{format_CO}
\alias{format_CO}
\title{Reformat the result of collect_CO().}
\usage{
format_CO(dict = NULL, orgdb = NULL)
}
\arguments{
\item{dict}{A result of collect_CO().}

\item{orgdb}{A genome annotation package.}
}
\value{
A formatted database.
}
\description{
This function reformats the result of collect_DO().
}
