% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{format_DO}
\alias{format_DO}
\title{Reformat a Disease Ontology database.}
\usage{
format_DO(dict, all_geneIDs, orgdb)
}
\arguments{
\item{dict}{A res slot of enrichDO().}

\item{all_geneIDs}{All genes in the form of ENTREZ ID.}

\item{orgdb}{A genome annotation package.}
}
\value{
A formatted database.
}
\description{
This function reformats a Disease Ontology database.
}
