% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{format_KEGG}
\alias{format_KEGG}
\title{Reformat collect_KEGG().}
\usage{
format_KEGG(dict = NULL, orgdb = NULL)
}
\arguments{
\item{dict}{A result of collect_KEGG().}

\item{orgdb}{A genome annotation package.}
}
\value{
A formatted database.
}
\description{
This function reformats collect_KEGG().
}
