% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_counts}
\alias{pbmc_counts}
\title{Single-cell RNA-seq data of peripheral blood mononuclear cells (PBMC 4k)}
\format{
Read count table
\describe{
  Gene-by-cell matrix data.
}
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression}
}
\usage{
pbmc_counts
}
\description{
A scRNA-seq dataset processed by conservative quality control.
The original data "4k PBMCs from a Healthy Donor" was downloaded from 10x
Genomics database.
}
\keyword{datasets}
