% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dataframe3D_asrt}
\alias{plot_dataframe3D_asrt}
\title{Visualize a three-dimensional data with labels and colors.}
\usage{
plot_dataframe3D_asrt(
  dataframe3D,
  labels = NULL,
  colors = NULL,
  theta = 30,
  phi = 30,
  title = "",
  title_size = 1.5,
  xlabel = "",
  ylabel = "",
  zlabel = ""
)
}
\arguments{
\item{dataframe3D}{A dataframe with three columns.}

\item{labels}{NULL or a vector of labels of all the samples,
corresponding to colors.}

\item{colors}{NULL or a vector of colors of all the samples,
corresponding to labels.}

\item{theta}{Angle of the plot.}

\item{phi}{Angle of the plot.}

\item{title}{Title.}

\item{title_size}{Title size.}

\item{xlabel}{x-axis label.}

\item{ylabel}{y-axis label.}

\item{zlabel}{z-axis label.}
}
\value{
A scatter3D object in plot3D package.
}
\description{
This function visualizes a three-dimensional data with labels and colors.
}
