% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_multiheatmaps}
\alias{plot_multiheatmaps}
\title{Visualize a three-dimensional data with labels and colors.}
\usage{
plot_multiheatmaps(
  ssm_list,
  gem_list = NULL,
  ssmlabel_list = NULL,
  gemlabel_list = NULL,
  nSamples = NULL,
  show_row_names = FALSE,
  title,
  title_size = 20
)
}
\arguments{
\item{ssm_list}{A list of sign-by-sample matrices.}

\item{gem_list}{A list of gene-by-sample matrices.}

\item{ssmlabel_list}{NULL or a list of dataframes of sample (cell)
labels and colors.
The length of the list must be as same as that of ssm_list, and
the order of labels in each list must be as same as those in ssm_list.}

\item{gemlabel_list}{NULL or a list of dataframes of sample (cell)
annotations and colors.
The length of the list must be as same as that of gem_list, and
the order of labels in each list must be as same as those in gem_list.}

\item{nSamples}{Number of samples (cells) used for random sampling.}

\item{show_row_names}{TRUE or FALSE: if TRUE, row names are shown.}

\item{title}{Title.}

\item{title_size}{Title size.}
}
\value{
A ComplexHeatmap object.
}
\description{
This function visualizes a three-dimensional data with labels and colors.
}
