% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{remove_samples_asrt}
\alias{remove_samples_asrt}
\title{Remove samples based on expression profiles across variables.}
\usage{
remove_samples_asrt(
  sce,
  min_nReads = 0,
  max_nReads = 1e+10,
  min_nGenes = 0,
  max_nGenes = 1e+10,
  min_percMT = 0,
  max_percMT = 100
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{min_nReads}{A minimum threshold value of the number of reads.}

\item{max_nReads}{A maximum threshold value of the number of reads.}

\item{min_nGenes}{A minimum threshold value of the number of non-zero
expressed genes.}

\item{max_nGenes}{A maximum threshold value of the number of non-zero
expressed genes.}

\item{min_percMT}{A minimum threshold value of the percent of reads that map
to mitochondrial genes.}

\item{max_percMT}{A maximum threshold value of the percent of reads that map
to mitochondrial genes.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function removes sample data by setting minimum and maximum threshold
  values for the metadata.
}
