% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signs.R
\name{remove_signs_manually}
\alias{remove_signs_manually}
\title{Remove signs by specifying keywords.}
\usage{
remove_signs_manually(sce, keywords = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{keywords}{keywords separated by pipes `|`.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function removes signs by specifying keywords.
}
