% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signs.R
\name{remove_signs_redundant}
\alias{remove_signs_redundant}
\title{Remove redundant signs using semantic similarity matrices.}
\usage{
remove_signs_redundant(
  sce = NULL,
  similarity_matrix = NULL,
  threshold = NULL,
  keep_rareID = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{similarity_matrix}{A semantic similarity matrix.}

\item{threshold}{A threshold value of semantic similarity, used for
regarding biological terms as similar ones}

\item{keep_rareID}{If TRUE, biological terms with the larger ICs are kept.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function removes redundant signs using semantic similarity matrices.
}
