% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signs.R
\name{select_signs_manually}
\alias{select_signs_manually}
\title{Select signs by specifying keywords.}
\usage{
select_signs_manually(sce, keywords = NULL)
}
\arguments{
\item{sce}{An ASURAT object.}

\item{keywords}{Keywords separated by a pipe.}
}
\value{
An ASURAT object.
}
\description{
This function selects signs by specifying keywords.
}
