% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_pass}
\alias{swap_pass}
\title{Perform one-shot adjacent swapping for each element.}
\usage{
swap_pass(data)
}
\arguments{
\item{data}{A list of vector and integer.}
}
\description{
Perform one-shot adjacent swapping for each element.
}
