% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASURAT.R
\docType{package}
\name{ASURAT}
\alias{ASURAT}
\title{Functional annotation-driven unsupervised clustering of SingleCell data.}
\description{
ASURAT is a software for single-cell data analysis.
   Using ASURAT, one can simultaneously perform unsupervised clustering and
   biological interpretation in terms of cell type, disease, biological
   process, and signaling pathway activity.
   Inputting a single-cell RNA-seq data and knowledge-based databases,
   such as Cell Ontology, Gene Ontology, KEGG, etc., ASURAT transforms
   gene expression tables into original multivariate tables, termed
   sign-by-sample matrices (SSMs).
}
