% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Perform bubble sorting, counting the number of steps.}
\usage{
bubble_sort(data)
}
\arguments{
\item{data}{A list of vector and integer.}
}
\description{
Perform bubble sorting, counting the number of steps.
}
