% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_signs.R
\docType{data}
\name{compute_nswap_001}
\alias{compute_nswap_001}
\title{Compute number of swaps.}
\format{
An object of class \code{character} of length 1.
}
\usage{
compute_nswap_001
}
\arguments{
\item{enco}{A vector.}
}
\description{
This function computes the number of swaps.
}
\keyword{datasets}
